this version compiles and produces a library and an executable for clustering
